Add HOC and remove location.reload
