+++ .github/workflows/cicd.yml
@@ -307,7 +307,6 @@ jobs:
     name: Push to registry
     needs: [version, unit_test_backend, unit_test_frontend, e2e_test, e2e_multi_windows_test]
     runs-on: ubuntu-22.04
-    if: needs.version.outputs.IS_RELEASE && !contains(github.ref, 'dependabot')
     env:
       IS_RELEASE: ${{ needs.version.outputs.IS_RELEASE }}
       ACTIONS_ALLOW_UNSECURE_COMMANDS: true
