+++ frontend/src/hooks/authorization/useGetUserAuthorization.ts
@@ -1,4 +1,3 @@
-import { isCypress } from '@utils/isCypress'
 import { useEffect, useState } from 'react'
 
 import { getCurrentUserAuthorization } from '../../domain/use_cases/authorization/getCurrentUserAuthorization'
@@ -9,30 +8,13 @@ import type { UserAuthorization } from '../../domain/entities/authorization/type
  * Get user authorization
  */
 export function useGetUserAuthorization(): UserAuthorization | undefined {
-  const IS_OIDC_ENABLED = isCypress() || import.meta.env.FRONTEND_OIDC_ENABLED === 'true'
   const [userAuthorization, setUserAuthorization] = useState<UserAuthorization | undefined>(undefined)
 
   useEffect(() => {
-    if (!IS_OIDC_ENABLED) {
-      /**
-       * This is used to have backward compatibility with the Apache .htacess authentication (on `/` and `/ext`) while the authentication
-       * is not yet activated, as the app is only protected by the entrypoint path.
-       */
-      const isExtPage = window.location.pathname === '/ext' || window.location.pathname === '/light'
-
-      setUserAuthorization({
-        isLogged: true,
-        isSuperUser: !isExtPage,
-        mustReload: false
-      })
-
-      return
-    }
-
     getCurrentUserAuthorization().then(nextUserAuthorization => {
       setUserAuthorization(nextUserAuthorization)
     })
-  }, [IS_OIDC_ENABLED])
+  }, [])
 
   return userAuthorization
 }
