+++ frontend/src/index.tsx
@@ -4,6 +4,7 @@ import { createRoot } from 'react-dom/client'
 import { AuthProvider, withAuth } from 'react-oidc-context'
 
 import { App } from './App'
+import { withAppAuthCurried } from './auth/appWithAuth'
 import { getOIDCConfig } from './auth/getOIDCConfig'
 
 import 'rsuite/dist/rsuite.min.css'
@@ -37,7 +38,7 @@ const root = createRoot(container)
 const { IS_OIDC_ENABLED, oidcConfig } = getOIDCConfig()
 
 if (IS_OIDC_ENABLED) {
-  const AppWithAuth = withAuth(App)
+  const AppWithAuth = withAuth(withAppAuthCurried()(App))
 
   root.render(
     // eslint-disable-next-line react/jsx-props-no-spreading
